Create UnitsPrintComponent to have same level of abstraction
